+++ inst/include/mvnorm.h
@@ -11,7 +11,9 @@ double dmvnorm(
 
 double pmvnorm(
     arma::vec const& x, arma::vec const& mean, arma::mat const& Sigma,
-    double abseps = 1e-3
+    double abseps = 1e-3,
+    Rcpp::Nullable<Rcpp::NumericVector> lower = R_NilValue,
+    std::string method = "genz", int draws = 500
 );
 
 arma::vec rmvnorm(