+++ src/internal.h
@@ -31,7 +31,9 @@ double dmvnorm_cpp(
 
 double pmvnorm_cpp(
     arma::vec const& x, arma::vec const& mean, arma::mat const& Sigma,
-    double abseps = 1e-3
+    double abseps = 1e-3,
+    Rcpp::Nullable<Rcpp::NumericVector> lower = R_NilValue,
+    std::string method = "genz", int draws = 500
 );
 
 arma::vec rmvnorm_cpp(